requirements.txt
django<1.9
django-debug-toolbar
djangorestframework
suds-jurko
gunicorn